requirements.txt
asn1crypto==0.24.0
backports-abc==0.5
bunch==1.0.1
CacheControl==0.12.5
cachetools==3.1.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.5
enum34==1.1.6
firebase-admin==2.15.1
future==0.17.1
futures==3.2.0
geographiclib==1.49
geopy==1.18.1
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-core==0.29.1
google-cloud-firestore==0.31.0
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.8
grpcio==1.18.0
idna==2.8
ipaddress==1.0.22
msgpack==0.6.1
protobuf==3.6.1
pur==5.2.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
python-dateutil==2.8.0
python-dotenv==0.10.1
python-telegram-bot==12.0.0b1
pytz==2018.9
requests==2.21.0
rsa==4.0
singledispatch==3.4.0.3
six==1.12.0
tornado==5.1.1
urllib3==1.24.1
